Allow for providing polyline styling options for multiple routes/tracks (closes #123)
This commit changes the expected type of the `polyline_options`
parameter from an object to an array of objects describing the desired
styling of each route/track in the GPX file. Backwards compatibility is
preserved if only one object is passed, making it apply to all
routes/tracks in the file as before.